+++ src/data/v2/bars.rs
@@ -177,7 +177,10 @@ Endpoint! {
   ],
   Err => GetError, [
     /// A query parameter was invalid.
-    /* 422 */ UNPROCESSABLE_ENTITY => InvalidInput,
+    /* 400 */ BAD_REQUEST => InvalidInput,
+    /// The request was not permitted. Possible reasons include usage of
+    /// the SIP feed without having the corresponding subscription.
+    /* 403 */ FORBIDDEN => NotPermitted,
   ]
 
   fn base_url() -> Option<Str> {
@@ -390,6 +393,30 @@ mod tests {
     assert_eq!(bars[0].close.to_u64(), Some(42));
     assert_eq!(bars[0].high.to_u64(), Some(43));
     assert_eq!(bars[0].low.to_u64(), Some(42));
+  }
+
+  /// Verify that we can specify the SIP feed as the data source to use.
+  #[test(tokio::test)]
+  async fn sip_feed() {
+    let api_info = ApiInfo::from_env().unwrap();
+    let client = Client::new(api_info);
+    let start = DateTime::from_str("2018-12-03T21:47:00Z").unwrap();
+    let end = DateTime::from_str("2018-12-07T21:47:00Z").unwrap();
+    let request = BarsReqInit {
+      limit: Some(2),
+      feed: Some(Feed::SIP),
+      ..Default::default()
+    }
+    .init("AAPL", start, end, TimeFrame::OneDay);
+
+    let result = client.issue::<Get>(&request).await;
+    // Unfortunately we can't really know whether the user has the
+    // unlimited plan and can access the SIP feed. So really all we can
+    // do here is accept both possible outcomes.
+    match result {
+      Ok(_) | Err(RequestError::Endpoint(GetError::NotPermitted(_))) => (),
+      err => panic!("Received unexpected error: {err:?}"),
+    }
   }
 
   /// Check that we fail as expected when an invalid page token is